+++ app/Tabletop/Battle/Exceptions/BattleDeedDenied.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GreatMarketrealmTabletop\Tabletop\Battle\Exceptions;
+
+use DomainException;
+
+defined('ABSPATH') || exit;
+
+final class BattleDeedDenied extends DomainException
+{
+}